+++ crates/matrix-sdk-ui/src/timeline/event_item/mod.rs
@@ -80,8 +80,14 @@ pub struct EventTimelineItem {
     pub(super) forwarder_profile: Option<TimelineDetails<Profile>>,
     /// The timestamp of the event.
     pub(super) timestamp: MilliSecondsSinceUnixEpoch,
-    /// The content of the event.
+    /// The content of the event. Might be redacted if a redaction for this
+    /// event is currently being sent or has been received from the server.
     pub(super) content: TimelineItemContent,
+    /// If a redaction for this event is currently being sent but the server
+    /// hasn't yet acknowledged it via its remote echo, the data
+    /// before redaction. This applies to all sorts of timeline items, including
+    /// state events. If no redaction is in flight, None.
+    pub(super) unredacted_item: Option<UnredactedEventTimelineItem>,
     /// The kind of event timeline item, local or remote.
     pub(super) kind: EventTimelineItemKind,
     /// Whether or not the event belongs to an encrypted room.
@@ -118,6 +124,20 @@ pub(crate) enum TimelineItemHandle<'a> {
     Local(&'a SendHandle),
 }
 
+/// A container for temporarily holding onto data that is going to be erased by
+/// a redaction once the server plays it back.
+#[derive(Clone, Debug)]
+pub(super) struct UnredactedEventTimelineItem {
+    /// The original content before redaction.
+    content: TimelineItemContent,
+
+    /// JSON of the original event.
+    pub(crate) original_json: Option<Raw<AnySyncTimelineEvent>>,
+
+    /// JSON of the latest edit to this item.
+    pub(crate) latest_edit_json: Option<Raw<AnySyncTimelineEvent>>,
+}
+
 impl EventTimelineItem {
     #[allow(clippy::too_many_arguments)]
     pub(super) fn new(
@@ -137,6 +157,7 @@ impl EventTimelineItem {
             forwarder_profile,
             timestamp,
             content,
+            unredacted_item: None,
             kind,
             is_room_encrypted,
         }
@@ -508,7 +529,12 @@ impl EventTimelineItem {
     }
 
     /// Create a clone of the current item, with content that's been redacted.
-    pub(super) fn redact(&self, rules: &RedactionRules) -> Self {
+    pub(super) fn redact(&self, rules: &RedactionRules, is_local: bool) -> Self {
+        let unredacted_item = is_local.then(|| UnredactedEventTimelineItem {
+            content: self.content.clone(),
+            original_json: self.original_json().cloned(),
+            latest_edit_json: self.latest_edit_json().cloned(),
+        });
         let content = self.content.redact(rules);
         let kind = match &self.kind {
             EventTimelineItemKind::Local(l) => EventTimelineItemKind::Local(l.clone()),
@@ -521,6 +547,35 @@ impl EventTimelineItem {
             forwarder_profile: self.forwarder_profile.clone(),
             timestamp: self.timestamp,
             content,
+            unredacted_item,
+            kind,
+            is_room_encrypted: self.is_room_encrypted,
+        }
+    }
+
+    /// Create a clone of the current item, with data restored from the
+    /// item's unredacted_item field (if it was previously set by a call to
+    /// the `redact(...)` method).
+    pub(super) fn unredact(&self) -> Self {
+        let Some(unredacted_item) = &self.unredacted_item else { return self.clone() };
+        let kind = match &self.kind {
+            EventTimelineItemKind::Local(l) => EventTimelineItemKind::Local(l.clone()),
+            EventTimelineItemKind::Remote(r) => {
+                EventTimelineItemKind::Remote(RemoteEventTimelineItem {
+                    original_json: unredacted_item.original_json.clone(),
+                    latest_edit_json: unredacted_item.latest_edit_json.clone(),
+                    ..r.clone()
+                })
+            }
+        };
+        Self {
+            sender: self.sender.clone(),
+            sender_profile: self.sender_profile.clone(),
+            forwarder: self.forwarder.clone(),
+            forwarder_profile: self.forwarder_profile.clone(),
+            timestamp: self.timestamp,
+            content: unredacted_item.content.clone(),
+            unredacted_item: None,
             kind,
             is_room_encrypted: self.is_room_encrypted,
         }