+++ crates/matrix-sdk/src/event_cache/tasks.rs
@@ -313,6 +313,11 @@ async fn handle_thread_subscriber_send_queue_update(
                     // Nothing to do, reactions don't count as a thread
                     // subscription.
                 }
+
+                LocalEchoContent::Redaction { .. } => {
+                    // Nothing to do, redactions don't count as a thread
+                    // subscription.
+                }
             }
             return true;
         }